LayoutTests/fast/canvas/philip/tests/2d.imageData.object.wrap.html

<!DOCTYPE html>
<title>Canvas test: 2d.imageData.object.wrap</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<link rel="prev" href="2d.imageData.object.string.html" title="2d.imageData.object.string">
<link rel="next" href="2d.imageData.object.round.html" title="2d.imageData.object.round">
<body class="show_output">
<p>
 <a href="2d.imageData.object.string.html" accesskey="p" title="[p] 2d.imageData.object.string">&lt;</a>
 <a href="index.html">[index]</a>
 <a href="2d.imageData.object.round.html" accesskey="n" title="[n] 2d.imageData.object.round">&gt;</a>
<h1><a href="index.2d.html">2d</a>.<a href="index.2d.imageData.html">imageData</a>.<a href="index.2d.imageData.object.html">object</a>.wrap</h1>
<p class="desc">ImageData.data wraps numbers modulo 256</p>
<div class="refs">References:
<ul>
<li><a href="spec.html#testrefs.2d.pixelarray.modify">2d.pixelarray.modify</a>

</ul>
</div>

<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

var imgdata = ctx.getImageData(0, 0, 10, 10);

imgdata.data[0] = 0;
_assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");
imgdata.data[0] = 300;
_assertSame(imgdata.data[0], 44, "imgdata.data[\""+(0)+"\"]", "44");
imgdata.data[0] = -100;
_assertSame(imgdata.data[0], 156, "imgdata.data[\""+(0)+"\"]", "156");

imgdata.data[0] = 200+Math.pow(2, 32);
_assertSame(imgdata.data[0], 200, "imgdata.data[\""+(0)+"\"]", "200");
imgdata.data[0] = -200-Math.pow(2, 32);
_assertSame(imgdata.data[0], 56, "imgdata.data[\""+(0)+"\"]", "56");

imgdata.data[0] = -Infinity;
_assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");
imgdata.data[0] = Infinity;
_assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");


});
</script>
