LayoutTests/platform/chromium-mac/canvas/philip/tests/2d.imageData.object.wrap-expected.txt

Failed assertion imgdata.data["0"] === 44 (got 255[number], expected 44[number])
Failed assertion imgdata.data["0"] === 156 (got 0[number], expected 156[number])
Failed assertion imgdata.data["0"] === 200 (got 255[number], expected 200[number])
Failed assertion imgdata.data["0"] === 56 (got 0[number], expected 56[number])
Failed assertion imgdata.data["0"] === 0 (got 255[number], expected 0[number])
